factor log functions into their own module
