+++ include/stumpless/level/notice.h
@@ -31,6 +31,7 @@
 
 #  ifndef STUMPLESS_DISABLE_NOTICE_LEVEL
 #    include <stumpless/config.h>
+#    include <stumpless/log.h>
 #    include <stumpless/severity.h>
 #    include <stumpless/target.h>
 #  endif
