+++ include/stumpless/level/trace.h
@@ -35,6 +35,7 @@
 
 #  ifndef STUMPLESS_DISABLE_TRACE_LEVEL
 #    include <stumpless/config.h>
+#    include <stumpless/log.h>
 #    include <stumpless/severity.h>
 #    include <stumpless/target.h>
 #  endif
