+++ include/stumpless/level/warning.h
@@ -31,6 +31,7 @@
 
 #  ifndef STUMPLESS_DISABLE_WARNING_LEVEL
 #    include <stumpless/config.h>
+#    include <stumpless/log.h>
 #    include <stumpless/severity.h>
 #    include <stumpless/target.h>
 #  endif
