+++ tools/check_headers/stumpless.yml
@@ -172,7 +172,7 @@
 "struct stumpless_version": "stumpless/version.h"
 "struct wel_data": "private/config/wel_supported.h"
 "struct wel_target": "private/target/wel.h"
-"stump": "stumpless/target.h"
+"stump": "stumpless/log.h"
 "stump_i": "stumpless/level/info.h"
 "stump_i_str": "stumpless/level/info.h"
 "stump_i_entry": "stumpless/level/info.h"
@@ -180,13 +180,13 @@
 "stump_i_log_str": "stumpless/level/info.h"
 "stump_i_message": "stumpless/level/info.h"
 "stump_i_message_str": "stumpless/level/info.h"
-"stump_str": "stumpless/target.h"
+"stump_str": "stumpless/log.h"
 "stump_t": "stumpless/level/trace.h"
 "stump_t_entry": "stumpless/level/trace.h"
 "stump_t_log": "stumpless/level/trace.h"
 "stump_t_message": "stumpless/level/trace.h"
-"stump_trace": "stumpless/target.h"
-"stump_trace_str": "stumpless/target.h"
+"stump_trace": "stumpless/log.h"
+"stump_trace_str": "stumpless/log.h"
 "stumpless_add_entry": "stumpless/target.h"
 "stumpless_add_log": "stumpless/target.h"
 "stumpless_add_log_str": "stumpless/target.h"
@@ -400,7 +400,7 @@
 "STUMPLESS_WINDOWS_EVENT_LOG_TARGETS_SUPPORTED": "stumpless/config.h"
 "STUMPLESS_WINDOWS_EVENT_LOG_CLOSE_FAILURE": "stumpless/error.h"
 "STUMPLESS_WINDOWS_EVENT_LOG_OPEN_FAILURE": "stumpless/error.h"
-"stumplog": "stumpless/target.h"
+"stumplog": "stumpless/log.h"
 "stumplog_a": "stumpless/level/alert.h"
 "stumplog_a_str": "stumpless/level/alert.h"
 "stumplog_c": "stumpless/level/crit.h"
@@ -415,12 +415,12 @@
 "stumplog_i_str": "stumpless/level/info.h"
 "stumplog_n": "stumpless/level/notice.h"
 "stumplog_n_str": "stumpless/level/notice.h"
-"stumplog_set_mask": "stumpless/target.h"
-"stumplog_str": "stumpless/target.h"
+"stumplog_set_mask": "stumpless/log.h"
+"stumplog_str": "stumpless/log.h"
 "stumplog_t": "stumpless/level/trace.h"
 "stumplog_t_str": "stumpless/level/trace.h"
-"stumplog_trace": "stumpless/target.h"
-"stumplog_trace_str": "stumpless/target.h"
+"stumplog_trace": "stumpless/log.h"
+"stumplog_trace_str": "stumpless/log.h"
 "stumplog_w": "stumpless/level/warning.h"
 "stumplog_w_str": "stumpless/level/warning.h"
 "sys_socket_close_network_target": "private/config/have_sys_socket.h"
@@ -445,8 +445,8 @@
 "unistd_getpid": "private/config/have_unistd.h"
 "unsupported_target_is_open": "private/target.h"
 "vsnprintf_s_format_string": "private/config/have_vsnprintf_s.h"
-"vstump": "stumpless/target.h"
-"vstump_trace": "stumpless/target.h"
+"vstump": "stumpless/log.h"
+"vstump_trace": "stumpless/log.h"
 "vstumpless_add_log": "stumpless/target.h"
 "vstumpless_add_message": "stumpless/target.h"
 "vstumpless_new_entry": "stumpless/entry.h"
@@ -454,8 +454,8 @@
 "vstumpless_set_wel_insertion_strings": "stumpless/config/wel_supported.h"
 "vstumpless_trace_log": "stumpless/target.h"
 "vstumpless_trace_message": "stumpless/target.h"
-"vstumplog": "stumpless/target.h"
-"vstumplog_trace": "stumpless/target.h"
+"vstumplog": "stumpless/log.h"
+"vstumplog_trace": "stumpless/log.h"
 "wel_open_default_target": "private/config/wel_supported.h"
 "windows_getpagesize": "private/config/have_windows.h"
 "windows_getpid": "private/config/have_windows.h"
